Move beego files in web
