--- conf/app.conf
@@ -1,3 +1,0 @@
-appname = litchi
-httpport = 8080
-runmode = dev
