--- main.go
@@ -1,10 +1,0 @@
-package main
-
-import (
-	"github.com/astaxie/beego"
-	_ "github.com/longan/litchi/routers"
-)
-
-func main() {
-	beego.Run()
-}
