--- routers/router.go
@@ -1,10 +1,0 @@
-package routers
-
-import (
-	"github.com/astaxie/beego"
-	"github.com/longan/litchi/controllers"
-)
-
-func init() {
-	beego.Router("/", &controllers.MainController{})
-}
