+++ web/conf/app.conf
@@ -1,0 +1,3 @@
+appname = litchi
+httpport = 8080
+runmode = dev
