+++ web/main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"github.com/astaxie/beego"
+	_ "github.com/longan/litchi/web/routers"
+)
+
+func main() {
+	beego.Run()
+}
